IGNITE-26371 Simplify
